refactor: Restructures repo & adds 'make run' cmd
Script files are now in src/awsweepbytag and 'make run' is used to
execute the script. README usage is also modified to reflect to simpler
command to run script.
