+++ src/awsweepbytag/get_and_order.py
@@ -4,8 +4,8 @@ import time
 import boto3
 import botocore.exceptions
 
-import get_other_ids
-import text_formatting as tf
+from awsweepbytag import get_other_ids
+from awsweepbytag import text_formatting as tf
 
 
 def get_resources_by_tag(tag_key: str, tag_value: str, regions: list[str]) -> list[dict[str, str]]:
